Added Self Delete Code.
The delete_self has to be set to true in the this_plugin section of the config. If not will not delete itself

Only activated during remote run. after everything else has been done and the logs ready to be sent out

Will append that last log result to the rest of the logs.

* Changed log class to only return one log

* changed plugins delete to always return that one log, it will return that in success or failing. But in rare cases where deleting itself causing issues (the pages should be cached for this run after deleting the source on disk), then exception info will be returned

* whatever is returned is appended to the run results. An absense of the self delete flag or it being turned off is also noted